requirements.txt
aniso8601
autopep8
django-cors-headers
django-graphql-jwt
graphene
graphene-django
graphql-core
graphql-relay
promise
pycodestyle
PyJWT
Rx
setuptools
singledispatch
wheel
asgiref==3.2.7
dj-database-url==0.5.0
Django==3.0.5
gunicorn==20.0.4
psycopg2-binary==2.8.4
pytz==2019.3
six==1.14.0
sqlparse==0.3.1
whitenoise==5.0.1